cleaned up command line options: make input and seed take BigIntegers while other INTEGER options remain taking regular int; fixed a bug in assertion failure report that didn't report the certainty correctly.
git-svn-id: svn://vsl.cis.udel.edu/civl/trunk@2919 fb995dde-84ed-4084-dfe6-e5aef3e2452c

+++ src/edu/udel/cis/vsl/civl/transform/common/GeneralWorker.java
@@ -1,5 +1,6 @@
 package edu.udel.cis.vsl.civl.transform.common;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -176,8 +177,9 @@ public class GeneralWorker extends BaseWorker {
 				&& inputVariables.containsKey(CIVL_argc_name)) {
 			Object CIVL_argc_obj = inputVariables.get(CIVL_argc_name);
 
-			assert CIVL_argc_obj instanceof Integer;
-			return this.integerConstant((int) CIVL_argc_obj);
+			assert CIVL_argc_obj instanceof BigInteger;
+			return this
+					.integerConstant(((BigInteger) CIVL_argc_obj).intValue());
 		}
 		return this.integerConstant(DEFAULT_ARGV_SIZE);
 	}
